+++ src/Endpoints/Databases.php
@@ -3,12 +3,13 @@
 namespace Vdhicts\Cyberfusion\ClusterApi\Endpoints;
 
 use Vdhicts\Cyberfusion\ClusterApi\Exceptions\RequestException;
-use Vdhicts\Cyberfusion\ClusterApi\Models\Cms;
+use Vdhicts\Cyberfusion\ClusterApi\Models\Database;
+use Vdhicts\Cyberfusion\ClusterApi\Models\DatabaseUsage;
 use Vdhicts\Cyberfusion\ClusterApi\Request;
 use Vdhicts\Cyberfusion\ClusterApi\Response;
 use Vdhicts\Cyberfusion\ClusterApi\Support\ListFilter;
 
-class Cmses extends Endpoint
+class Databases extends Endpoint
 {
     /**
      * @param ListFilter|null $filter
@@ -23,7 +24,7 @@ class Cmses extends Endpoint
 
         $request = (new Request())
             ->setMethod(Request::METHOD_GET)
-            ->setUrl(sprintf('cmses/?%s', http_build_query($filter->toArray())));
+            ->setUrl(sprintf('databases/?%s', http_build_query($filter->toArray())));
 
         $response = $this
             ->client
@@ -33,9 +34,9 @@ class Cmses extends Endpoint
         }
 
         return $response->setData([
-            'cmses' => array_map(
+            'databases' => array_map(
                 function (array $data) {
-                    return (new Cms())->fromArray($data);
+                    return (new Database())->fromArray($data);
                 },
                 $response->getData()
             ),
@@ -51,7 +52,7 @@ class Cmses extends Endpoint
     {
         $request = (new Request())
             ->setMethod(Request::METHOD_GET)
-            ->setUrl(sprintf('cmses/%d', $id));
+            ->setUrl(sprintf('databases/%d', $id));
 
         $response = $this
             ->client
@@ -61,7 +62,30 @@ class Cmses extends Endpoint
         }
 
         return $response->setData([
-            'cms' => (new Cms())->fromArray($response->getData()),
+            'database' => (new Database())->fromArray($response->getData()),
+        ]);
+    }
+
+    /**
+     * @param int $id
+     * @return Response
+     * @throws RequestException
+     */
+    public function usages(int $id): Response
+    {
+        $request = (new Request())
+            ->setMethod(Request::METHOD_GET)
+            ->setUrl(sprintf('databases/usages/%d', $id));
+
+        $response = $this
+            ->client
+            ->request($request);
+        if (! $response->isSuccess()) {
+            return $response;
+        }
+
+        return $response->setData([
+            'databaseUsage' => (new DatabaseUsage())->fromArray($response->getData()),
         ]);
     }
 }