+++ src/Models/Cluster.php
@@ -9,6 +9,7 @@ class Cluster implements Model
 {
     public string $name;
     public ?string $group = null;
+    public ?string $dataDirectory = null;
     public ?int $id = null;
     public ?string $createdAt = null;
     public ?string $updatedAt = null;
@@ -18,6 +19,7 @@ class Cluster implements Model
         $cluster = new self();
         $cluster->name = Arr::get($data, 'name');
         $cluster->group = Arr::get($data, 'group');
+        $cluster->dataDirectory = Arr::get($data, 'data_directory');
         $cluster->id = Arr::get($data, 'id');
         $cluster->createdAt = Arr::get($data, 'created_at');
         $cluster->updatedAt = Arr::get($data, 'updated_at');
@@ -29,6 +31,7 @@ class Cluster implements Model
         return [
             'name' => $this->name,
             'group' => $this->group,
+            'data_directory' => $this->dataDirectory,
             'id' => $this->id,
             'created_at' => $this->createdAt,
             'updated_at' => $this->updatedAt,