+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0]
+
+### Added
+
+- Add sandbox mode. Enable the sandbox mode in the configuration class, see the readme for more information.
+- Add the `databases` endpoint, which allows you to list and get databases and view the disk usage.
+- Add the `data_directory` attribute to the cluster model.
+- Add psalm as static analysis tool.
+
 ## [1.0.0]
 
 ### Added